+++ src/main/java/com/beforehairshop/demo/member/repository/MemberRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, BigInteger> {
 
     Optional<Member> findOneByEmailAndStatusIsLessThan(String email, int status);
 }
